登録者本人以外の更新・削除を制御する機能の追加

+++ app/api/item/delete/[id]/route.ts
@@ -3,13 +3,21 @@ import connectDB from "@/app/utils/database";
 import { ItemModel } from "@/app/utils/schemaModels";
 
 export async function DELETE(request: NextRequest, context: NextFetchEvent) {
+  const reqBody = await request.json(); // リクエストボディから更新データを取得
   try {
     await connectDB();
     const id = request.nextUrl.pathname.split("/").pop();
-    await ItemModel.deleteOne({ _id: id });
-    return NextResponse.json({
-      message: "アイテムの削除に成功しました",
-    });
+    const currentItem = await ItemModel.findById(id); // 現在のアイテムを取得
+    if (currentItem.email !== reqBody.email) {
+      await ItemModel.deleteOne({ _id: id });
+      return NextResponse.json({
+        message: "アイテムの削除に成功しました",
+      });
+    } else {
+      return NextResponse.json({
+        message: "他のユーザーのアイテムは削除できません",
+      });
+    }
   } catch (err) {
     return NextResponse.json({ message: "アイテムの削除に失敗しました" });
   }
